requirements.txt
absl-py==2.1.0
aiohttp==3.9.3
aiorun==2023.7.2
aiosignal==1.3.1
alabaster==0.7.13
alembic==1.13.1
altair==5.2.0
annotated-types==0.6.0
appdirs==1.4.4
appnope==0.1.3
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
autograd==1.6.2
av==12.0.0
Babel==2.13.1
backcall==0.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
bs4==0.0.2
cachetools==5.3.2
category-encoders==2.6.3
certifi==2024.2.2
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorlog==6.8.2
comm==0.1.4
contourpy==1.1.1
cycler==0.12.1
datasets==2.17.1
debugpy==1.8.0
decorator==4.4.2
dill==0.3.8
docker-pycreds==0.4.0
docutils==0.20.1
exceptiongroup==1.1.3
executing==2.0.0
Farama-Notifications==0.0.4
fastjsonschema==2.18.1
ffmpeg==1.4
fica==0.3.1
filelock==3.13.1
fonttools==4.43.1
frozenlist==1.4.1
fsspec==2023.10.0
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.42
glob2==0.7
google-api-core==2.12.0
google-api-python-client==2.106.0
google-auth==2.23.4
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
googletrans==3.0.0
graphviz==0.20.1
grpcio==1.60.1
gspread==5.12.0
gym==0.26.2
gym-notices==0.0.8
gymnasium==0.29.1
h11==0.9.0
h2==3.2.0
hl7==0.4.5
hpack==3.0.0
hstspreload==2024.2.1
httpcore==0.9.1
httplib2==0.22.0
httpx==0.13.3
huggingface-hub==0.20.3
hyperframe==5.2.0
idna==2.10
imageio==2.32.0
imagesize==1.4.1
importlib-metadata==7.0.1
iopath==0.1.10
ipykernel==6.25.2
ipython==8.16.1
ipywidgets==8.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter_client==8.4.0
jupyter_core==5.4.0
jupyterlab_widgets==3.0.10
jupytext==1.15.2
kiwisolver==1.4.5
libretranslatepy==2.1.1
lightning-utilities==0.11.2
lxml==5.1.0
Mako==1.3.2
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.1
matplotlib-inline==0.1.6
mdit-py-plugins==0.4.0
mdurl==0.1.2
memory-profiler==0.61.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
optuna==3.5.0
otter-grader==4.1.0
packaging==23.2
pandas==2.1.2
pandoc==2.3
parameterized==0.9.0
parso==0.8.3
patsy==0.5.6
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.1
platformdirs==3.11.0
plotly==5.18.0
plumbum==1.8.2
ply==3.11
portalocker==2.8.2
proglog==0.1.10
prometheus_client==0.20.0
prompt-toolkit==3.0.39
protobuf==4.25.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.0
pyarrow-hotfix==0.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
pygame==2.5.2
Pygments==2.16.1
pyparsing==3.1.1
python-dateutil==2.8.2
python-on-whales==0.66.0
pytorch-lightning==2.2.3
pytorchvideo==0.1.5
pytz==2023.3.post1
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rich==13.7.0
rpds-py==0.12.0
rsa==4.9
sacrebleu==1.5.0
safetensors==0.4.2
scikit-learn==1.3.2
scipy==1.11.3
seaborn==0.13.0
sentencepiece==0.2.0
sentry-sdk==1.40.5
seqeval==1.2.2
setproctitle==1.3.3
six==1.16.0
skorch==0.15.0
smmap==5.0.1
sniffio==1.3.0
snowballstemmer==2.2.0
soupsieve==2.5
Sphinx==7.2.6
sphinxcontrib-applehelp==1.0.7
sphinxcontrib-devhelp==1.0.5
sphinxcontrib-htmlhelp==2.0.4
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.6
sphinxcontrib-serializinghtml==1.1.9
SQLAlchemy==2.0.28
stack-data==0.6.3
statsmodels==0.14.1
streamlit==1.31.1
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
tensorboard==2.16.2
tensorboard-data-server==0.7.2
termcolor==2.4.0
textblob==0.18.0.post0
threadpoolctl==3.2.0
tokenizers==0.15.2
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torch_geometric==2.4.0
torchaudio==2.2.2
torchdata==0.7.1
torchmetrics==1.3.2
torchvision==0.17.2
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.38.1
translate==3.6.1
translator==0.0.9
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.0.7
validators==0.22.0
wandb==0.16.3
wcwidth==0.2.8
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.10
wrapt==1.15.0
xxhash==3.4.1
yacs==0.1.8
yarl==1.9.4
zipp==3.17.0
